Fetch account to include time entry metadata
If we fetch the account instead of the time entries, we get entries
for the last 10 days and all workspace/project/task names in a single
API call.

For a more generic/flexible solution I would make two API calls: one
for time entries, and another (optional) one for metadata. But I'll be
the only user for this receiver, so I can indulge.